requirements.txt
Flask>=2.1
python-dotenv>=1.0
requests>=2.28
gunicorn>=20.1